Re-work ptrace, simplify, remove need for multiple threads
This puts all the tracing into a single thread that handles all
the different processes being traced.  The simplification allows
follwing of fork/clone calls and makes it easier to track resources.

Since we're running now from a single thread we can remove the overhead
of mutexes that protect various data strcutures.

Signed-off-by: Colin Ian King <colin.king@canonical.com>