Check for new procs and add them to CPU stats
This also re-works the way CPU stat durations are calculated.
Originally the code assumed the duration was for the fixed length
of the life of health check, but we can now have processes that start
during the test, so we need to account for the total life from start
to end.

Signed-off-by: Colin Ian King <colin.king@canonical.com>